+++ src/Cool/Gpu/OpenGL/Shader.cpp
@@ -10,7 +10,7 @@ static void assert_shader_is_bound(GLuint id)
 #if defined(DEBUG)
     GLint current_id;
     glGetIntegerv(GL_CURRENT_PROGRAM, &current_id);
-    assert(current_id == id && "The shader must be bound before setting any uniform.");
+    assert(static_cast<GLuint>(current_id) == id && "The shader must be bound before setting any uniform.");
 #endif
 }
