+++ src/Cool/String/String.h
@@ -25,10 +25,10 @@ void replace_all(std::string& str, std::string_view from, std::string_view to);
  * @param min_nb_of_characters The minimum number of characters that will be present in the output string.
  */
 template<typename T>
-std::string to_string(T val, int min_nb_of_characters = 0)
+std::string to_string(T val, size_t min_nb_of_characters = 0)
 {
     std::string str = std::to_string(val);
-    int         dn  = min_nb_of_characters - static_cast<int>(str.size());
+    size_t      dn  = min_nb_of_characters - str.size();
     if (dn > 0) {
         str = std::string(dn, '0') + str;
     }
